refactor rendering
