+++ src/llamafactory/v1/plugins/trainer_plugins/batching.py
@@ -19,7 +19,7 @@ from typing import Any
 import torch
 from torch.utils.data import default_collate
 
-from ...core.utils.rendering import pad_and_truncate
+from ...core.utils.collation import pad_and_truncate
 from ...utils.constants import IGNORE_INDEX
 from ...utils.objects import StatefulBuffer
 from ...utils.plugin import BasePlugin
